+++ src/konpy/tasks.clj
@@ -13,8 +13,8 @@
   "show this weeks assignments.
    this page must provide link to answer and views."
   [_]
-  (let [tasks-q '[:find ?week ?num ?task
-                  :keys week num task
+  (let [tasks-q '[:find ?e ?week ?num ?task
+                  :keys e week num task
                   :in $ ?week
                   :where
                   [?e :week ?week]
@@ -24,12 +24,12 @@
               (sort-by :num))]
     (page
       [:div
-       (for [{:keys [week num task]} ret]
+       (for [{:keys [e week num task]} ret]
          [:div {:class "flex"}
           [:span (str week "-" num " " task)]
           [:span
            [:a {:class "rounded-xl text-white bg-sky-500 hover:bg-sky-700 active:bg-red-500"
-                :href "/answers"}
+                :href (str "/answer/" e)}
             "回答"]]])])))
 
 (defn tasks-all